organized app hooks
